// frontend/src/api/web-api-client.ts
export class UserDetailDto implements IUserDetailDto {
    id?: number;
    firstName?: string;
    lastName?: string;
    email?: string;
    instrument?: string;
    phone?: string;
    address?: string;
    zip?: string;
    city?: string;
    roles?: string[];
    isInitialPassword?: boolean;

    [key: string]: any;

    constructor(data?: IUserDetailDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.id = _data["id"];
            this.firstName = _data["firstName"];
            this.lastName = _data["lastName"];
            this.email = _data["email"];
            this.instrument = _data["instrument"];
            this.phone = _data["phone"];
            this.address = _data["address"];
            this.zip = _data["zip"];
            this.city = _data["city"];
            if (Array.isArray(_data["roles"])) {
                this.roles = [] as any;
                for (let item of _data["roles"])
                    this.roles!.push(item);
            }
            this.isInitialPassword = _data["isInitialPassword"];
        }
    }

    static fromJS(data: any): UserDetailDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserDetailDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["id"] = this.id;
        data["firstName"] = this.firstName;
        data["lastName"] = this.lastName;
        data["email"] = this.email;
        data["instrument"] = this.instrument;
        data["phone"] = this.phone;
        data["address"] = this.address;
        data["zip"] = this.zip;
        data["city"] = this.city;
        if (Array.isArray(this.roles)) {
            data["roles"] = [];
            for (let item of this.roles)
                data["roles"].push(item);
        }
        data["isInitialPassword"] = this.isInitialPassword;
        return data;
    }
}

export interface IUserDetailDto {
    id?: number;
    firstName?: string;
    lastName?: string;
    email?: string;
    instrument?: string;
    phone?: string;
    address?: string;
    zip?: string;
    city?: string;
    roles?: string[];
    isInitialPassword?: boolean;

    [key: string]: any;
}

export class EventDto implements IEventDto {
    id?: number;
    title?: string;
    description?: string;
    date?: Date;
    location?: string;
    type?: string;
    attendanceStatus?: string;

    [key: string]: any;

    constructor(data?: IEventDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.id = _data["id"];
            this.title = _data["title"];
            this.description = _data["description"];
            this.date = _data["date"] ? new Date(_data["date"].toString()) : undefined as any;
            this.location = _data["location"];
            this.type = _data["type"];
            this.attendanceStatus = _data["attendanceStatus"];
        }
    }

    static fromJS(data: any): EventDto {
        data = typeof data === 'object' ? data : {};
        let result = new EventDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["id"] = this.id;
        data["title"] = this.title;
        data["description"] = this.description;
        data["date"] = this.date ? this.date.toISOString() : undefined as any;
        data["location"] = this.location;
        data["type"] = this.type;
        data["attendanceStatus"] = this.attendanceStatus;
        return data;
    }
}

export interface IEventDto {
    id?: number;
    title?: string;
    description?: string;
    date?: Date;
    location?: string;
    type?: string;
    attendanceStatus?: string;

    [key: string]: any;
}
// ...
export class AttendanceDto implements IAttendanceDto {
    userId?: number;
    user?: UserDetailDto;
    status?: string;
    comment?: string;

    [key: string]: any;

    constructor(data?: IAttendanceDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.userId = _data["userId"];
            this.user = _data["user"] ? UserDetailDto.fromJS(_data["user"]) : undefined as any;
            this.status = _data["status"];
            this.comment = _data["comment"];
        }
    }

    static fromJS(data: any): AttendanceDto {
        data = typeof data === 'object' ? data : {};
        let result = new AttendanceDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["userId"] = this.userId;
        data["user"] = this.user ? this.user.toJSON() : undefined as any;
        data["status"] = this.status;
        data["comment"] = this.comment;
        return data;
    }
}

export interface IAttendanceDto {
    userId?: number;
    user?: UserDetailDto;
    status?: string;
    comment?: string;

    [key: string]: any;
}